Make "Pooling" section in README more understandable (#64)
I tried to understand the "pooling" part in the English README, but I did not really understand most of the sentences. So I used [DeepL](www.deepl.com) to translate the Chinese version, and I finally understood. So I thought maybe it would be a good idea to submit the DeepL translation to the README, as I think it is far more understandable.